# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FolioOut": {
        "properties": {
          "balance": {
            "title": "Balance",
            "type": "integer"
          },
          "charges_total": {
            "title": "Charges Total",
            "type": "integer"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "guest_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FolioItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "opened_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "payments_total": {
            "title": "Payments Total",
            "type": "integer"
          },
          "reservation_room_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Room Id"
          },
          "status": {
            "$ref": "#/components/schemas/FolioStatus"
          },
          "type": {
            "$ref": "#/components/schemas/FolioType"
          }
        },
        "required": [
          "id",
          "number",
          "type",
          "status",
          "guest_id",
          "reservation_room_id",
          "charges_total",
          "payments_total",
          "balance",
          "opened_at",
          "closed_at",
          "items"
        ],
        "title": "FolioOut",
        "type": "object"
      },
      "FolioStatus": {
        "enum": [
          "OPEN",
          "CLOSED",
          "SETTLED"
        ],
        "title": "FolioStatus",
        "type": "string"
      },
      "FolioType": {
        "enum": [
          "GUEST",
          "MASTER",
          "WALK_IN",
          "TABLE"
        ],
        "title": "FolioType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvoiceOut": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "folio_id": {
            "format": "uuid",
            "title": "Folio Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_provisional": {
            "title": "Is Provisional",
            "type": "boolean"
          },
          "issued_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineOut"
            },
            "title": "Lines",
            "type": "array"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "subtotal": {
            "title": "Subtotal",
            "type": "integer"
          },
          "tax_total": {
            "title": "Tax Total",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "is_provisional",
          "folio_id",
          "status",
          "issued_at",
          "subtotal",
          "tax_total",
          "total",
          "currency",
          "lines"
        ],
        "title": "InvoiceOut",
        "type": "object"
      },
...
      "PaymentIn": {
        "properties": {
          "amount": {
            "description": "FCFA",
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "type": "integer"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UUID v7 genere par la tablette ; absent = genere par le serveur",
            "title": "Id"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "required": [
          "method",
          "amount"
        ],
        "title": "PaymentIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Atrium",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/folios": {
      "get": {
        "operationId": "list_folios_api_v1_folios_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FolioStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "guest_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Guest Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FolioOut"
                  },
                  "title": "Response List Folios Api V1 Folios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Folios",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}": {
      "get": {
        "operationId": "get_folio_api_v1_folios__folio_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Folio",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}/close": {
      "post": {
        "operationId": "close_folio_api_v1_folios__folio_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Close Folio",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}/invoice": {
      "post": {
        "description": "Gel du folio a l'instant present (F1.4).\n\nCette route s'execute directement sur le serveur central -- contrairement\na une facture emise hors ligne depuis une tablette, elle n'a pas besoin de\nnumero provisoire : `next_number` attribue tout de suite le numero\nlegal definitif (voir le commentaire du modele `Invoice` et\napp/services/numbering.py).",
        "operationId": "issue_invoice_api_v1_folios__folio_id__invoice_post",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Facture deja emise pour ce folio : celle-la"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Issue Invoice",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}/items": {
      "post": {
        "description": "Charge manuelle (minibar, blanchisserie, remise...).\n\nUne remise (`category=DISCOUNT`) exige en plus `folio.discount` -- c'est\nexactement l'exemple cite par docs/01-modele-de-donnees.md pour justifier\ndes permissions granulaires plutot qu'un seul controle global sur le folio.",
        "operationId": "add_folio_item_api_v1_folios__folio_id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolioItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Charge deja enregistree (meme id) : etat actuel"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolioItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Folio Item",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}/payments": {
      "post": {
        "description": "Encaissement. Un paiement rejoue (meme id) n'est jamais compte deux\n\nfois : 200 avec l'etat actuel du folio, sans rien reecrire.",
        "operationId": "record_payment_api_v1_folios__folio_id__payments_post",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Paiement deja enregistre (meme id) : etat du folio"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Record Payment",
        "tags": [
          "facturation"
        ]
      }
    },
    "/api/v1/folios/{folio_id}/post-stay-nights": {
      "post": {
        "description": "Porte au folio toutes les nuits encore non postees du sejour associe\n\n(`StayNight.is_posted`) -- version simplifiee, folio par folio, de la\ncloture journaliere du paragraphe 5.1, qui operera plus tard sur tout\nl'hotel d'un coup.",
        "operationId": "post_stay_nights_api_v1_folios__folio_id__post_stay_nights_post",
        "parameters": [
          {
            "in": "path",
            "name": "folio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Folio Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Post Stay Nights",
        "tags": [
          "facturation"
        ]
      }
    },
...
  }
}